AI/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
colorama==0.4.6
Cython==0.29.32
Django==3.2.13
django-cors-headers==3.14.0
djangorestframework==3.14.0
docker==6.0.1
filelock==3.9.0
fst-pso==1.8.1
FuzzyTM==2.0.5
gluonnlp==0.10.0
graphviz==0.8.4
huggingface-hub==0.13.0
idna==2.6
Jinja2==3.1.2
kobert-transformers==0.4.1
MarkupSafe==2.1.2
miniful==0.0.6
mpmath==1.3.0
mxnet==1.7.0.post2
networkx==3.0
numpy==1.19.5
packaging==23.0
pandas==1.5.3
pyFUME==0.2.25
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scipy==1.10.1
sentencepiece==0.1.97
simpful==2.10.0
six==1.16.0
smart-open==6.3.0
sqlparse==0.4.3
sympy==1.11.1
tokenizers==0.13.2
torch==2.0.0
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.15
websocket-client==1.5.1
